scim_user: allow to ignore certain attributes during updates
this is useful e.g. when you want to allow users to keep their custom
names while only updating the username and job title on Slack

also updated Slack example make use of that functionality, and also
added the `"displayName"` attribute there which is apparently required
for usernames.  who knew!